% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br_mt_1_8K_9classes_6bands}
\alias{br_mt_1_8K_9classes_6bands}
\title{Time series of agriculture and forestry in Mato Grosso from MOD13Q1}
\format{
A \code{sits} tibble with 1.8K samples and 9 classes ("Cerrado",
  "Fallow_Cotton", "Forest", "Millet_Cotton", "Pasture",
  "Soy_Corn", "Soy_Cotton", "Soy_Fallow", "Soy_Millet").
   Each time series comprehends 12 months (23 data points)
   from MOD13Q1 product, with 6 bands.
}
\usage{
data(br_mt_1_8K_9classes_6bands)
}
\description{
A dataset containing a sits tibble with time series samples from Brazilian
Mato Grosso State (Amazon and Cerrado biomes).
}
\keyword{datasets}
