% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_cerrado_cbers}
\alias{samples_cerrado_cbers}
\title{Samples for Brazilian Cerrado using CBERS AWFI image}
\format{
A tibble with 1 rows and 7 variables: longitude, latitude, start_date
  (initial date of the time series), end_date (final date of the time series),
  label (the class label associated to the sample),
  cube (the name of the data cube associated with the time series),
  time_series (list containing a tibble with the values of the time series).
  Each time series comprehends one year of 16-day composites
  from AWFI images, comprising 23 data points and 6 bands
  (blue, green, red, nir, ndvi, evi)
}
\usage{
data(cbers_samples_022024)
}
\description{
A dataset containing a tibble with time series sampled on the
Brazilian Cerrado. The time series come from a set of CBERS-4 AWFI images
over tile "022024" of the Brazilian Data Cube.
The CBERS-4 AWFI is a sensor with 64 meter resolution.
}
\keyword{datasets}
