% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_cerrado_lc8}
\alias{samples_cerrado_lc8}
\title{Time series of agriculture and forestry in Cerrado biome from 
BDC Landsat 8 16-days composite (stack) data cube}
\format{
A \code{sits} tibble with 67.3K samples and 8 classes ("Annual_Crop",
"Forest", "Nat_Non_Vegetated", "Pasture", "Savanna",
"Semi_Perennial_Crop", "Silviculture", "Water").
Each time series comprehends 12 months (24 data points)
from BDC Landsat 8 16-days composite (stack) data cube, with 9 bands,
with cloud values removed and interpolated using sits. 
   
This dataset is composed from a variety of sources:
Field work of TerraClass-Cerrado Team (EMBRAPA and INPE);
Field work of Pastagem.org Team (LAPIG-UFG);
Image interpretation of BDC Team (INPE)
Automatic extraction and cross comparison analysis among 
Pastagem.org (LAPIG-UFG), PRODES-Cerrado (INPE), 
and MapBiomas (MapBiomas) maps;
}
\usage{
data("samples_cerrado_lc8")
}
\description{
A dataset containing a sits tibble with time series samples 
from Brazilian all Cerrado biome extension.
}
\keyword{datasets}
