% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_cerrado_mapsamples_lc8}
\alias{samples_cerrado_mapsamples_lc8}
\title{Time series of agriculture and forestry in Cerrado biome from 
BDC Landsat 8 16-days composite (stack) data cube}
\format{
A \code{sits} tibble with 48K samples and 10 classes ("Annual_Crop", 
"Cerradao", "Cerrado", "Nat_NonVeg", "Open_Cerrado", "Pasture", 
"Perennial_Crop", "Silviculture", "Sugarcane", "Water").

Each time series comprehends 12 months (24 data points)
from BDC Landsat 8 16-days composite (stack) data cube, with 2 vegetation
indices (NDVI and EVI) and 7 spectral bands (B1-7). All time series have
bee cloud observation removed and no data interpolated using sits. 
   
This dataset is composed from automatic extraction and cross comparison 
analysis among Pastagem.org (LAPIG-UFG), PRODES-Cerrado (INPE), 
and MapBiomas (MapBiomas) maps;
}
\usage{
data("samples_cerrado_mapsamples_lc8")
}
\description{
A dataset containing a sits tibble with time series samples 
from Brazilian all Cerrado biome extension from existing maps (see details).
}
\keyword{datasets}
