% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_cerrado_mod13q1}
\alias{samples_cerrado_mod13q1}
\title{Yearly samples of 14 classes on Cerrado biome from MOD13Q1 images}
\format{
A \code{sits} tibble with 61K samples. A \code{sits} tibble contains the following meta-data:
  longitude (East-west coordinate of the time series sample in WGS 84),
  latitude (North-south coordinate of the time series sample in WGS 84),
  start_date (initial date of the time series),
  end_date (final date of the time series),
  label (the class label associated to the sample),
  coverage (the name of the coverage associated with the data), and
  time_series ( list containing a tibble with the values of the time series).

  The samples are from many sources. It has 13 classes ("Araguaia", "Campo_Cerrado",
  "Cerradao", "Cerrado", "Cerrado_Rupestre", "Dunas",
  "Fallow_Cotton", "Millet_Cotton", "Pasture", "Silvicultura",
  "Soy_Corn", "Soy_Cotton", "Soy_Fallow", "Soy_Millet").

  Each time series comprehends 12 months (23 data points) from MOD13Q1 product, and has 4 bands
  ("evi", "ndvi", "mir", and "nir").
}
\usage{
data(cerrado_61K_14classes_4bands)
}
\description{
A dataset containing a sits tibble with time series samples from Brazilian Cerrado biome.
}
\keyword{datasets}
