% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_S2_T20LKP_2018_2019}
\alias{samples_S2_T20LKP_2018_2019}
\title{Samples for an area in Amazonia using SENTINEL-2/2A MSI data}
\format{
A tibble with 1 rows and 7 variables: longitude, latitude, start_date
  (initial date of the time series), end_date (final date of the time series),
  label (the class label associated to the sample),
  cube (the name of the data cube associated with the time series),
  time_series (list containing a tibble with the values of the time series).
  Each time series comprehends one year of composites of S2 and S2A images,
  comprising 36 data points and 11 bands and indices
  (B02, B03, B07, B08, B8A, B11, B12, evi, ndvi, ndmi, savi)
}
\usage{
data(samples_S2_T20LKP_2018_2019)
}
\description{
A dataset containing a tibble with time series sampled on the
Brazilian Amazonia (Rondonia state). The time series come from a set of SENTINEL-2/2A MSI images
over tile "T20LKP" of the standard S2 grid.
}
\keyword{datasets}
