% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_validation_cerrado}
\alias{samples_validation_cerrado}
\title{LULC Cerrado classification validation}
\usage{
data("samples_validation_cerrado")
}
\description{
This dataset provides samples used for validation of the Cerrado 
LULC classification (see above). The authors did a systematic sampling of 
the Cerrado biome using a 20 x 20 km grid with a total of 5402 points. 
These samples are independent of the training set used in the classification. 
They were interpreted by five specialists using high resolution 
images from the same period of the classification. 
This resulted in 5286 evaluation samples thus distributed: 
"Annual Crop" (553), "Cerrado" (3155), "Natural Non Vegetated" (44), 
"Pasture" (1246), "Perennial Crop" (38), "Silviculture" (94), 
"Sugarcane" (77), and "Water" (79).  

Reference: Simoes, Rolf, Gilberto Camara, Gilberto Queiroz, 
Felipe Souza, Pedro R. Andrade, Lorena Santos, Alexandre Carvalho, 
and Karine Ferreira. 2021. 
"Satellite Image Time Series Analysis for Big Earth Observation Data". 
Remote Sensing 13, no. 13: 2428. https://doi.org/10.3390/rs13132428.
}
\keyword{datasets}
