% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_cerrado_lc8}
\alias{samples_cerrado_lc8}
\title{Samples of Cerrado biome used in the paper "Satellite Image Time 
Series Analysis for Big Earth ObservationData"}
\format{
A \code{sits} tibble with ~48K samples and 10 classes ("Annual_Crop", 
"Cerradao", "Cerrado", "Nat_NonVeg", "Open_Cerrado", "Pasture", 
"Perennial_Crop", "Silviculture", "Sugarcane", "Water").
}
\usage{
data("samples_cerrado_lc8")
}
\description{
A sits tibble with time series samples from Brazilian Cerrado biome.
The samples is based on a systematic sampling design of 5-kilometers
over all Cerrado extension. 

Labels were automatically attributed from
a cross-comparison among land cover products of the year of 2018 from
the following sources: Pastagem.org (LAPIG-UFG), IBGE Land use and land 
cover map (IBGE), and MapBiomas (MapBiomas). 

The classification labels 
were: "Annual_Crop", "Cerradao", "Cerrado", "Nat_NonVeg", "Open_Cerrado", 
"Pasture", "Perennial_Crop", "Silviculture", "Sugarcane", and "Water". 
All those points not consistent with these classes as well
as those with conflicting labels among input maps were discarded.

The time series were extracted from the Landsat 8 BDC data cube 
(LC8_30_16D_STK-1). These time series comprehends a period of 12 months 
(24 observations) from "2017-09-01" to "2018-08-31".
The extracted attributes were: NDVI, EVI, B1, B2, B3, B4, B5, B6, and B7.
Cloudy values were removed and interpolated.

The validation data set used in the referred paper is in the external 
data folder of this package that can be accessed using:

\code{system.file("extdata/cerrado/validation.csv", package = "sitsdata")}

This validation samples as well as the training data set were collected 
as part of the Brazil Data Cube project (http://brazildatacube.org).
More details on how these data sets were obtained can be found in the 
referred paper.
}
\keyword{datasets}
